+++ src/pages/Main/Main.jsx
@@ -1,4 +1,4 @@
-import Header from "../components/Header/Header";
+import Header from "../../components/Header/Header";
 import {
   Content,
   H2,
