--- src/pages/MessageData.js
@@ -1,266 +1,0 @@
-export const messageData = [
-  {
-    recipientId: 1,
-    sender: "김동훈",
-    profileImageURL:
-      "https://fastly.picsum.photos/id/311/200/200.jpg?hmac=CHiYGYQ3Xpesshw5eYWH7U0Kyl9zMTZLQuRDU4OtyH8",
-
-    relationship: "가족",
-    content:
-      "코로나가 또다시 기승을 부리는 요즘이네요.건강, 체력 모두 조심 또 하세요!",
-    font: "Pretendard",
-    createdAt: "2025-02-06",
-  },
-  {
-    recipientId: 1,
-    sender: "김훈",
-    profileImageURL:
-      "https://fastly.picsum.photos/id/311/200/200.jpg?hmac=CHiYGYQ3Xpesshw5eYWH7U0Kyl9zMTZLQuRDU4OtyH8",
-
-    relationship: "가족",
-    content:
-      "코로나가 또다시 기승을 부리는 요즘이네요.건강, 체력 모두 조심 또 하세요!",
-    font: "Pretendard",
-    createdAt: "2025-02-06",
-  },
-  {
-    recipientId: 1,
-    sender: "동훈",
-    profileImageURL:
-      "https://fastly.picsum.photos/id/311/200/200.jpg?hmac=CHiYGYQ3Xpesshw5eYWH7U0Kyl9zMTZLQuRDU4OtyH8",
-
-    relationship: "가족",
-    content:
-      "코로나가 또다시 기승을 부리는 요즘이네요.건강, 체력 모두 조심 또 하세요!",
-    font: "Pretendard",
-    createdAt: "2025-02-06",
-  },
-  {
-    recipientId: 1,
-    sender: "김훈",
-    profileImageURL:
-      "https://fastly.picsum.photos/id/311/200/200.jpg?hmac=CHiYGYQ3Xpesshw5eYWH7U0Kyl9zMTZLQuRDU4OtyH8",
-
-    relationship: "가족",
-    content:
-      "코로나가 또다시 기승을 부리는 요즘이네요.건강, 체력 모두 조심 또 하세요!",
-    font: "Pretendard",
-    createdAt: "2025-02-06",
-  },
-  {
-    recipientId: 1,
-    sender: "동훈",
-    profileImageURL:
-      "https://fastly.picsum.photos/id/311/200/200.jpg?hmac=CHiYGYQ3Xpesshw5eYWH7U0Kyl9zMTZLQuRDU4OtyH8",
-
-    relationship: "가족",
-    content:
-      "코로나가 또다시 기승을 부리는 요즘이네요.건강, 체력 모두 조심 또 하세요!",
-    font: "Pretendard",
-    createdAt: "2025-02-06",
-  },
-  {
-    recipientId: 1,
-    sender: "김훈",
-    profileImageURL:
-      "https://fastly.picsum.photos/id/311/200/200.jpg?hmac=CHiYGYQ3Xpesshw5eYWH7U0Kyl9zMTZLQuRDU4OtyH8",
-
-    relationship: "가족",
-    content:
-      "코로나가 또다시 기승을 부리는 요즘이네요.건강, 체력 모두 조심 또 하세요!",
-    font: "Pretendard",
-    createdAt: "2015-02-06",
-  },
-  {
-    recipientId: 1,
-    sender: "동훈",
-    profileImageURL:
-      "https://fastly.picsum.photos/id/311/200/200.jpg?hmac=CHiYGYQ3Xpesshw5eYWH7U0Kyl9zMTZLQuRDU4OtyH8",
-
-    relationship: "가족",
-    content:
-      "코로나가 또다시 기승을 부리는 요즘이네요.건강, 체력 모두 조심 또 하세요!",
-    font: "Pretendard",
-    createdAt: "2025-01-06",
-  },
-  {
-    recipientId: 1,
-    sender: "김훈",
-    profileImageURL:
-      "https://fastly.picsum.photos/id/311/200/200.jpg?hmac=CHiYGYQ3Xpesshw5eYWH7U0Kyl9zMTZLQuRDU4OtyH8",
-
-    relationship: "가족",
-    content:
-      "코로나가 또다시 기승을 부리는 요즘이네요.건강, 체력 모두 조심 또 하세요!",
-    font: "Pretendard",
-    createdAt: "2025-02-06",
-  },
-  {
-    recipientId: 1,
-    sender: "동훈",
-    profileImageURL:
-      "https://fastly.picsum.photos/id/311/200/200.jpg?hmac=CHiYGYQ3Xpesshw5eYWH7U0Kyl9zMTZLQuRDU4OtyH8",
-
-    relationship: "가족",
-    content:
-      "코로나가 또다시 기승을 부리는 요즘이네요.건강, 체력 모두 조심 또 하세요!",
-    font: "Pretendard",
-    createdAt: "2025-02-06",
-  },
-  {
-    recipientId: 1,
-    sender: "김훈",
-    profileImageURL:
-      "https://fastly.picsum.photos/id/311/200/200.jpg?hmac=CHiYGYQ3Xpesshw5eYWH7U0Kyl9zMTZLQuRDU4OtyH8",
-
-    relationship: "가족",
-    content:
-      "코로나가 또다시 기승을 부리는 요즘이네요.건강, 체력 모두 조심 또 하세요!",
-    font: "Pretendard",
-    createdAt: "2021-02-06",
-  },
-  {
-    recipientId: 1,
-    sender: "동훈",
-    profileImageURL:
-      "https://fastly.picsum.photos/id/311/200/200.jpg?hmac=CHiYGYQ3Xpesshw5eYWH7U0Kyl9zMTZLQuRDU4OtyH8",
-
-    relationship: "가족",
-    content:
-      "코로나가 또다시 기승을 부리는 요즘이네요.건강, 체력 모두 조심 또 하세요!",
-    font: "Pretendard",
-    createdAt: "2025-02-06",
-  },
-  {
-    recipientId: 1,
-    sender: "김훈",
-    profileImageURL:
-      "https://fastly.picsum.photos/id/311/200/200.jpg?hmac=CHiYGYQ3Xpesshw5eYWH7U0Kyl9zMTZLQuRDU4OtyH8",
-
-    relationship: "가족",
-    content:
-      "코로나가 또다시 기승을 부리는 요즘이네요.건강, 체력 모두 조심 또 하세요!",
-    font: "Pretendard",
-    createdAt: "2025-02-05",
-  },
-  {
-    recipientId: 1,
-    sender: "동훈",
-    profileImageURL:
-      "https://fastly.picsum.photos/id/311/200/200.jpg?hmac=CHiYGYQ3Xpesshw5eYWH7U0Kyl9zMTZLQuRDU4OtyH8",
-
-    relationship: "가족",
-    content:
-      "코로나가 또다시 기승을 부리는 요즘이네요.건강, 체력 모두 조심 또 하세요!",
-    font: "Pretendard",
-    createdAt: "2029-02-06",
-  },
-  {
-    recipientId: 1,
-    sender: "동훈",
-    profileImageURL:
-      "https://fastly.picsum.photos/id/311/200/200.jpg?hmac=CHiYGYQ3Xpesshw5eYWH7U0Kyl9zMTZLQuRDU4OtyH8",
-
-    relationship: "가족",
-    content:
-      "코로나가 또다시 기승을 부리는 요즘이네요.건강, 체력 모두 조심 또 하세요!",
-    font: "Pretendard",
-    createdAt: "2025-02-06",
-  },
-  {
-    recipientId: 1,
-    sender: "김훈",
-    profileImageURL:
-      "https://fastly.picsum.photos/id/311/200/200.jpg?hmac=CHiYGYQ3Xpesshw5eYWH7U0Kyl9zMTZLQuRDU4OtyH8",
-
-    relationship: "가족",
-    content:
-      "코로나가 또다시 기승을 부리는 요즘이네요.건강, 체력 모두 조심 또 하세요!",
-    font: "Pretendard",
-    createdAt: "2025-02-05",
-  },
-  {
-    recipientId: 1,
-    sender: "동훈",
-    profileImageURL:
-      "https://fastly.picsum.photos/id/311/200/200.jpg?hmac=CHiYGYQ3Xpesshw5eYWH7U0Kyl9zMTZLQuRDU4OtyH8",
-
-    relationship: "가족",
-    content:
-      "코로나가 또다시 기승을 부리는 요즘이네요.건강, 체력 모두 조심 또 하세요!",
-    font: "Pretendard",
-    createdAt: "2029-02-06",
-  },
-  {
-    recipientId: 1,
-    sender: "동훈",
-    profileImageURL:
-      "https://fastly.picsum.photos/id/311/200/200.jpg?hmac=CHiYGYQ3Xpesshw5eYWH7U0Kyl9zMTZLQuRDU4OtyH8",
-
-    relationship: "가족",
-    content:
-      "코로나가 또다시 기승을 부리는 요즘이네요.건강, 체력 모두 조심 또 하세요!",
-    font: "Pretendard",
-    createdAt: "2025-02-06",
-  },
-  {
-    recipientId: 1,
-    sender: "김훈",
-    profileImageURL:
-      "https://fastly.picsum.photos/id/311/200/200.jpg?hmac=CHiYGYQ3Xpesshw5eYWH7U0Kyl9zMTZLQuRDU4OtyH8",
-
-    relationship: "가족",
-    content:
-      "코로나가 또다시 기승을 부리는 요즘이네요.건강, 체력 모두 조심 또 하세요!",
-    font: "Pretendard",
-    createdAt: "2025-02-05",
-  },
-  {
-    recipientId: 1,
-    sender: "동훈",
-    profileImageURL:
-      "https://fastly.picsum.photos/id/311/200/200.jpg?hmac=CHiYGYQ3Xpesshw5eYWH7U0Kyl9zMTZLQuRDU4OtyH8",
-
-    relationship: "가족",
-    content:
-      "코로나가 또다시 기승을 부리는 요즘이네요.건강, 체력 모두 조심 또 하세요!",
-    font: "Pretendard",
-    createdAt: "2029-02-06",
-  },
-  {
-    recipientId: 1,
-    sender: "동훈",
-    profileImageURL:
-      "https://fastly.picsum.photos/id/311/200/200.jpg?hmac=CHiYGYQ3Xpesshw5eYWH7U0Kyl9zMTZLQuRDU4OtyH8",
-
-    relationship: "가족",
-    content:
-      "코로나가 또다시 기승을 부리는 요즘이네요.건강, 체력 모두 조심 또 하세요!",
-    font: "Pretendard",
-    createdAt: "2025-02-06",
-  },
-  {
-    recipientId: 1,
-    sender: "김훈",
-    profileImageURL:
-      "https://fastly.picsum.photos/id/311/200/200.jpg?hmac=CHiYGYQ3Xpesshw5eYWH7U0Kyl9zMTZLQuRDU4OtyH8",
-
-    relationship: "가족",
-    content:
-      "코로나가 또다시 기승을 부리는 요즘이네요.건강, 체력 모두 조심 또 하세요!",
-    font: "Pretendard",
-    createdAt: "2025-02-05",
-  },
-  {
-    recipientId: 1,
-    sender: "동훈",
-    profileImageURL:
-      "https://fastly.picsum.photos/id/311/200/200.jpg?hmac=CHiYGYQ3Xpesshw5eYWH7U0Kyl9zMTZLQuRDU4OtyH8",
-
-    relationship: "가족",
-    content:
-      "코로나가 또다시 기승을 부리는 요즘이네요.건강, 체력 모두 조심 또 하세요!",
-    font: "Pretendard",
-    createdAt: "2029-02-06",
-  },
-];
